requirements.txt
scrapy==2.1.0
google-cloud-bigquery==1.25.0
apache-airflow[gcp]==1.10.10
